Calculator Ver 1 - some errors
Still some issues to work out; many buttons work, but the operators do
not. Currently the program treats the second input number as both
numbers. For example, if first input is 5 and second input is 2, it
assumes first input is also 2. Thus, first + second = 4, not 7. I will
look into this further.